+++ test/text.go
@@ -1,0 +1,77 @@
+package test
+
+import (
+	"encoding"
+	"fmt"
+
+	"github.com/stretchr/testify/assert"
+)
+
+// CaseText represents one specific test case for MarshalText and/or UnmarshalText method.
+type CaseText[T any] struct {
+	Constraint Constraint
+	Error      AssertErrorFunc
+	Data       string
+	Value      T
+}
+
+// MarshalText runs all passed test cases of method with same name.
+func MarshalText[T any](t TestingT, cases []CaseText[T]) {
+	t.Helper()
+
+	for i, c := range cases {
+		if i == 0 {
+			if _, ok := any(c.Value).(encoding.TextMarshaler); !ok {
+				assert.FailNowf(t, "unable to test MarshalText", "type %T must implements encoding.TextMarshaler", c.Value)
+				return
+			}
+		}
+
+		if !isForMarshal(c.Constraint) {
+			continue
+		}
+
+		failInfo := fmt.Sprintf("case %d failed", i)
+		b, err := any(c.Value).(encoding.TextMarshaler).MarshalText()
+		if c.Error != nil {
+			if c.Error(t, err, failInfo) {
+				assert.Nil(t, b, failInfo)
+			}
+		} else {
+			if assert.NoError(t, err, failInfo) {
+				assert.Equal(t, c.Data, string(b), failInfo)
+			}
+		}
+	}
+}
+
+// UnmarshalText runs all passed test cases of method with same name.
+func UnmarshalText[T any](t TestingT, cases []CaseText[T], helper TypeHelper[T]) {
+	t.Helper()
+
+	for i, c := range cases {
+		if i == 0 {
+			if _, ok := any(&c.Value).(encoding.TextUnmarshaler); !ok {
+				assert.FailNowf(t, "unable to test UnmarshalText", "type *%T must implements encoding.TextUnmarshaler", &c.Value)
+				return
+			}
+		}
+
+		if !isForUnmarshal(c.Constraint) {
+			continue
+		}
+
+		failInfo := fmt.Sprintf("case %d failed", i)
+		v := helperNew[T](helper, &c.Value)
+		err := any(v).(encoding.TextUnmarshaler).UnmarshalText([]byte(c.Data))
+		if c.Error != nil {
+			if c.Error(t, err, failInfo) {
+				helperAssertEmpty(helper, t, v, failInfo)
+			}
+		} else {
+			if assert.NoError(t, err, failInfo) {
+				helperAssertEqual(helper, t, &c.Value, v, failInfo)
+			}
+		}
+	}
+}